requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
dj-database-url
groq
faiss-cpu
sentence-transformers
python-dotenv
